requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-model-utils==4.0.0
djangorestframework==3.12.1
gunicorn==20.0.4
h11==0.11.0
httpcore==0.12.0
imutils==0.5.3
numpy==1.19.4
opencv-python==4.4.0.44
Pillow==8.0.1
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2020.4
sniffio==1.2.0
sqlparse==0.4.1
whitenoise==5.2.0
